polish pr according to Pu
