+++ dizoo/procgen/entry/coinrun_onppo_main.py
@@ -1,16 +1,19 @@
+import os
+from functools import partial
+
+import gym
+import numpy as np
+from easydict import EasyDict
+from tensorboardX import SummaryWriter
+
+from ding.torch_utils import to_ndarray
 from ding.worker import BaseLearner, SampleSerialCollector, InteractionSerialEvaluator
 from ding.model import VAC
 from ding.policy import PPOPolicy
 from ding.envs import DingEnvWrapper, FinalEvalRewardEnv, BaseEnvManager
 from ding.config import compile_config
 from ding.utils import set_pkg_seed
-
-from dizoo.dmc2gym.config.dmc2gym_ppo_config import cartpole_balance_ppo_config
-from tensorboardX import SummaryWriter
-import os
-import dmc2gym
-from dizoo.dmc2gym.envs.dmc2gym_env import *
-from easydict import EasyDict
+from dizoo.procgen.config.coinrun_ppo_config import coinrun_ppo_config
 
 
 class InfoWrapper(gym.Wrapper):
@@ -18,14 +21,11 @@ class InfoWrapper(gym.Wrapper):
         super().__init__(env)
         cfg = EasyDict(cfg)
         self._cfg = cfg
-        self._observation_space = dmc2gym_env_info[cfg.domain_name][cfg.task_name]["observation_space"](
-            from_pixels=self._cfg["from_pixels"],
-            height=self._cfg["height"],
-            width=self._cfg["width"],
-            channels_first=self._cfg["channels_first"]
+        self._observation_space = gym.spaces.Box(
+            low=np.zeros(shape=(3, 64, 64)), high=np.ones(shape=(3, 64, 64)) * 255, shape=(3, 64, 64), dtype=np.float32
         )
-        self._action_space = dmc2gym_env_info[cfg.domain_name][cfg.task_name]["action_space"]
-        self._reward_space = dmc2gym_env_info[cfg.domain_name][cfg.task_name]["reward_space"](self._cfg["frame_skip"])
+        self._action_space = gym.spaces.Discrete(15)
+        self._reward_space = gym.spaces.Box(low=float("-inf"), high=float("inf"), shape=(1, ), dtype=np.float32)
 
     def seed(self, seed: int, dynamic_seed: bool = True) -> None:
         self._seed = seed
@@ -33,45 +33,35 @@ class InfoWrapper(gym.Wrapper):
         np.random.seed(self._seed)
 
     def _process_obs(self, obs):
-        if self._cfg["from_pixels"]:
-            obs = to_ndarray(obs).astype(np.uint8)
-        else:
-            obs = to_ndarray(obs).astype(np.float32)
+        obs = to_ndarray(obs)
+        obs = np.transpose(obs, (2, 0, 1))
+        obs = obs.astype(np.float32)
         return obs
 
     def step(self, action):
-        action = np.array([action]).astype('float32')
         obs, reward, done, info = self.env.step(action)
-        return self._process_obs(obs), reward, done, info
+        return self._process_obs(obs), reward, bool(done), info
 
     def reset(self):
         obs = self.env.reset()
         return self._process_obs(obs)
 
 
-def wrapped_dmc2gym_env(cfg):
-    default_cfg = {
-            "frame_skip": 3,
-            "from_pixels": True,
-            "visualize_reward": False,
-            "height": 100,
-            "width": 100,
-            "channels_first": True,
-    }
+def wrapped_procgen_env(cfg):
+    default_cfg = dict(
+        control_level=True,
+        start_level=0,
+        num_levels=0,
+        env_id='coinrun',
+    )
     default_cfg.update(cfg)
-
+    default_cfg = EasyDict(default_cfg)
 
     return DingEnvWrapper(
-        dmc2gym.make(
-            domain_name=default_cfg["domain_name"],
-            task_name=default_cfg["task_name"],
-            seed=1,
-            visualize_reward=default_cfg["visualize_reward"],
-            from_pixels=default_cfg["from_pixels"],
-            height=default_cfg["height"],
-            width=default_cfg["width"],
-            frame_skip=default_cfg["frame_skip"]
-        )
+        gym.make('procgen:procgen-' + default_cfg.env_id + '-v0',
+                 start_level=default_cfg.start_level,
+                 num_levels=default_cfg.num_levels) if default_cfg.control_level else
+        gym.make('procgen:procgen-' + default_cfg.env_id + '-v0', start_level=0, num_levels=1)
         ,
         cfg={
             'env_wrapper': [
@@ -88,14 +78,10 @@ def main(cfg, seed=0, max_iterations=int(1e10)):
         save_cfg=True
     )
     collector_env_num, evaluator_env_num = cfg.env.collector_env_num, cfg.env.evaluator_env_num
-    collector_env = BaseEnvManager(env_fn=[lambda: wrapped_dmc2gym_env(cartpole_balance_ppo_config.env) for _ in range(collector_env_num)],
-                                   cfg=cfg.env.manager)
-    evaluator_env = BaseEnvManager(env_fn=[lambda: wrapped_dmc2gym_env(cartpole_balance_ppo_config.env) for _ in range(evaluator_env_num)],
-                                   cfg=cfg.env.manager)
-    # collector_env = BaseEnvManager(env_fn=[lambda: DMC2GymEnv(cfg.env) for _ in range(collector_env_num)],
-    #                                cfg=cfg.env.manager)
-    # evaluator_env = BaseEnvManager(env_fn=[lambda: DMC2GymEnv(cfg.env) for _ in range(evaluator_env_num)],
-    #                                cfg=cfg.env.manager)
+    collector_env = BaseEnvManager(env_fn=[partial(wrapped_procgen_env, cfg=coinrun_ppo_config.env)
+                                           for _ in range(collector_env_num)], cfg=cfg.env.manager)
+    evaluator_env = BaseEnvManager(env_fn=[partial(wrapped_procgen_env, cfg=coinrun_ppo_config.env)
+                                           for _ in range(evaluator_env_num)], cfg=cfg.env.manager)
 
     collector_env.seed(seed)
     evaluator_env.seed(seed, dynamic_seed=False)
@@ -122,4 +108,4 @@ def main(cfg, seed=0, max_iterations=int(1e10)):
 
 
 if __name__ == '__main__':
-    main(cartpole_balance_ppo_config)
+    main(coinrun_ppo_default_config)
